addressed requests2
